feat: :sparkles: búsqueda de productos y ajustes en el crud del usuario
se validan los datos del crud de usuarios, en la búsqueda de productos se toman los terminos de la consulta y se realiza el query con cada termino relevante

+++ src/Crud/Users/Delete.php
@@ -30,6 +30,14 @@ class Delete
                 throw new \Exception('Invalid user');
             }
             $connection = Database::connect();
+            $sql = 'SELECT * FROM users WHERE id = :id LIMIT 1';
+            $statement = $connection->prepare($sql);
+            $statement->bindParam('id', $id, \PDO::PARAM_INT);
+            $statement->execute();
+            $user = $statement->fetch(\PDO::FETCH_ASSOC);
+            if (empty($user)) {
+                throw new \Exception('User not found');
+            }
             $sql = 'DELETE FROM users WHERE id = :id';
             $statement = $connection->prepare($sql);
             $statement->bindParam('id', $id, \PDO::PARAM_INT);
